capture-jobcard/requirements.txt
typer>=0.9.0
rich>=13.0.0
requests>=2.31.0
pymupdf>=1.24.0
Pillow>=10.0.0
google-auth>=2.20.0
